+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `runAsyncFunction` for custom async functions. (#862)
 - Added `CellError` type to `getNamedExpressionValue` return types. (#862)
 
+- Added support for inferring return types based on their values in custom function plugins. (#875)
+
 ## [1.3.0] - 2021-10-20
 
 ### Added